requirements.txt
spidev
numpy
scipy
plotly
python-box
influxdb-client
tomli
tomli_w
prettytable
ruff
